removed Avatar.ios
